:boom: Move children generation callback into control_bar
From there, it cannot reference the variable `app`, but we likely do not need this and the callback can become just a function so we will deal with this later.
